+++ x/oracle/abci/preblocker_test.go
@@ -1,0 +1,5 @@
+package abci_test
+
+func (s *IntegrationTestSuite) TestPreBlocker() {
+
+}
