+++ x/oracle/abci/proposal_test.go
@@ -1,0 +1,9 @@
+package abci_test
+
+func (s *IntegrationTestSuite) TestPrepareProposalHandler() {
+
+}
+
+func (s *IntegrationTestSuite) TestProcessProposalHandler() {
+
+}
