+++ x/oracle/abci/voteextension_test.go
@@ -1,0 +1,9 @@
+package abci_test
+
+func (s *IntegrationTestSuite) TestExtendVoteHandler() {
+
+}
+
+func (s *IntegrationTestSuite) TestVerifyVoteExtensionHandler() {
+
+}
